move the default form value logic to main component
